Move services to the right place
